requirements.txt
transformers
torch
speechrecognition
pyttsx3
python-dotenv
